Splits MainFrame events into its own file
